Always create main venv with python3
Tricky, because this venv is shared by multiple tasks, so any one of
them could actually create the venv (whichever runs first, if it doesn't
exist).

NB - If the virtualenv already exists, but was created with Python2, it
MUST be deleted manually before (re)deployment. Otherwise it won't be
re-created by Ansible.